add client, refactor news
