+++ lib/graphql/client.dart
@@ -1,0 +1,54 @@
+import 'package:http/http.dart' as http;
+import 'news.dart';
+import 'dart:convert';
+import 'dart:async';
+
+String _url = 'https://techconf.mercari.com/2018/api/query';
+
+class Client {
+  static final Client _singleton = new Client._internal();
+
+  http.Client _client;
+
+  factory Client() {
+    _singleton._client = http.Client();
+    return _singleton;
+  }
+
+  Client._internal();
+
+  Future<List<News>> fetchNews() async {
+    String _query = """
+  {
+    newsList {
+      nodes {
+        id
+        date
+        message
+        messageJa
+        link
+      }
+    }
+  }
+""";
+    var response = await _client.post(_url,
+        body: json.encode({
+          'query': _query,
+        }));
+    if (response.statusCode == 200) {
+      var decoded = json.decode(utf8.decode(response.bodyBytes));
+      List<dynamic> nodes = decoded['data']['newsList']['nodes'];
+      return nodes.map((n) => News.fromJson(n)).toList();
+    } else {
+      // エラーが起きたらデフォルトのやつを表示しておく
+      return [
+        News(
+            date: '2018-09-04',
+            id: '1',
+            link: 'https://techconf.mercari.com/2018',
+            messageJa: 'Mercari Tech Conf 2018のWebページが公開されました。',
+            message: 'Mercari Tech Conf 2018のWebページが公開されました。'),
+      ];
+    }
+  }
+}
